nanpresance/requirements.txt
colorama==0.4.3
Django==2.2.5
django-qr-code==1.1.0
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0
pillow
djangorestframework
drf-dynamic-fields